+++ amplipi/app.py
@@ -87,6 +87,7 @@ logger.setLevel(logging.DEBUG)
 sh = logging.StreamHandler(sys.stdout)
 logger.addHandler(sh)
 
+
 class SimplifyingRouter(APIRouter):
   """
   Overrides the route decorator logic to:
@@ -500,9 +501,11 @@ def exec_command(cmd: models.StreamCommand, ctrl: Api = Depends(get_ctrl), sid: 
     * Like/Love Current Song: **love**
     * Ban Current Song (pandora only): **ban**
     * Shelve Current Song (pandora only): **shelve**
+    * Restart Stream **restart**
 
   Supported commands are reported in an attached stream's info.stream_cmds"""
   return code_response(ctrl, ctrl.exec_stream_command(sid, cmd=cmd))
+
 
 @api.get('/api/streams/{sid}/browse', tags=['stream'])
 def browse_stream(ctrl: Api = Depends(get_ctrl), sid: int = params.StreamID) -> models.BrowsableItemResponse:
@@ -517,7 +520,7 @@ def browse_stream(ctrl: Api = Depends(get_ctrl), sid: int = params.StreamID) -> 
 
 
 @api.get('/api/streams/{sid}/{pid}/browse', tags=['stream'])
-def browse_stream_child(ctrl: Api = Depends(get_ctrl), sid: int = params.StreamID, pid : int = params.ParentID) -> models.BrowsableItemResponse:
+def browse_stream_child(ctrl: Api = Depends(get_ctrl), sid: int = params.StreamID, pid: int = params.ParentID) -> models.BrowsableItemResponse:
   """ Browse the children of a media item in the current stream """
   stream = ctrl.streams[sid]
   if stream is None:
@@ -527,8 +530,9 @@ def browse_stream_child(ctrl: Api = Depends(get_ctrl), sid: int = params.StreamI
 
   return models.BrowsableItemResponse(items=stream.browse(parent=pid))
 
+
 @api.get('/api/streams/{sid}/{cid}/play', tags=['stream'])
-def play_stream_child(ctrl: Api = Depends(get_ctrl), sid: int = params.StreamID, cid : int = params.ChildID) -> models.Status:
+def play_stream_child(ctrl: Api = Depends(get_ctrl), sid: int = params.StreamID, cid: int = params.ChildID) -> models.Status:
   """ Play a child item from the current stream """
   stream = ctrl.streams[sid]
   if stream is None: