+++ amplipi/display/__init__.py
@@ -1,2 +1,2 @@
 """ Display """
-__all__ = ["common", "display", "einkdisplay", "epd2in13_V3", "tftdisplay"]
+__all__ = ["common", "display", "einkdisplay", "epd2in13_V3", "tftdisplay", "statusinterface"]