+++ amplipi/streams.py
@@ -145,6 +145,17 @@ class BaseStream:
     self.state = 'connected'
     self.src = src
 
+  def restart(self):
+    """Reset this stream by disconnecting and reconnecting"""
+    try:
+      self.send_cmd('stop')
+    except:
+      logger.info(f'Stream {self.name} does not have a stop response')
+    last_src = self.src # Disconnect sets self.src to none, so temp variable used to keep track
+    self.disconnect()
+    time.sleep(0.1)
+    self.connect(last_src)
+
   def is_connected(self) -> bool:
     return self.src is not None
 
@@ -294,6 +305,16 @@ class PersistentStream(BaseStream):
     if self.is_activated():
       self.deactivate()
       time.sleep(0.1)  # wait a bit just in case
+
+  def restart(self):
+    """Reset this stream by disconnecting and reconnecting"""
+    try:
+      self.send_cmd('stop')
+    except:
+      logger.info(f'Stream {self.name} does not have a stop response')
+    self.deactivate()
+    time.sleep(0.1)
+    self.activate()
 
   def connect(self, src: int):
     """ Connect an output to a given audio source """
@@ -1541,16 +1562,25 @@ class LMS(PersistentStream):
         src_config_folder = f'{utils.get_folder("config")}/srcs/v{self.vsrc}'
         os.system(f'rm -f {src_config_folder}')
         self.proc.terminate()
-        self.proc.communicate()
-        if self.meta_proc is not None:
-          self.meta_proc.terminate()
-          self.meta_proc.communicate()
-          self.meta_proc = None
+        self.proc.communicate(timeout=10)
       except Exception as e:
-        logger.exception(f"failed to terminate LMS stream {self.name}: {e} \nforcefully killing")
-        self.proc.kill()
-        self.proc.communicate()
+        logger.exception(f"failed to gracefully terminate LMS stream {self.name}: {e}")
+        logger.warning(f"forcefully killing LMS stream {self.name}")
+        os.killpg(self.proc.pid, signal.SIGKILL)
+        self.proc.communicate(timeout=3)
+
+    if self.meta_proc is not None:
+      try:
+        self.meta_proc.terminate()
+        self.meta_proc.communicate(timeout=10)
+      except:
+        logger.exception(f"failed to gracefully terminate LMS meta proc for {self.name}: {e}")
+        logger.warning(f"forcefully killing LMS meta proc for {self.name}")
+        os.killpg(self.meta_proc.pid, signal.SIGKILL)
+        self.meta_proc.communicate(timeout=3)
+
     self.proc = None
+    self.meta_proc = None
 
   def info(self) -> models.SourceInfo:
     # Opens and reads the metadata.json file every time the info def is called